backend/requirements.txt
fastapi
uvicorn[standard]
httpx
polyline
python-dotenv
